+++ packages/donor/src/screens/UpcommingDonation/UpcomingDonationScreen.tsx
@@ -5,7 +5,6 @@ import {
   Hospital,
   LinkUtils,
   LocaleUtils,
-  DeviceUtils,
 } from "@zm-blood-components/common";
 import styles from "./UpcomingDonationScreen.module.scss";
 import ZMScreen from "../../components/basic/ZMScreen";
@@ -16,9 +15,9 @@ import Cancellation from "../../assets/images/cancelation.svg";
 import Whatsapp from "../../assets/images/whatsup-color-big.svg";
 import TrashIcon from "../../assets/icons/trash.svg";
 import UpcomingDonationInfo from "./UpcomingDonationInfo";
-import ICalendarLink from "react-icalendar-link";
+import EventCalendarLink from "./EventCalendarLink";
 
-const EVENT_DURATION = 90;
+const EVENT_DURATION_BY_MINUTES = 90;
 
 export interface UpcomingDonationScreenProps {
   bookedAppointment: BookedAppointment;
@@ -31,45 +30,6 @@ export default function UpcomingDonationScreen({
   onCancel,
   bookedAppointment,
 }: UpcomingDonationScreenProps) {
-  const eventDateStart = new Date(bookedAppointment.donationStartTimeMillis);
-  const eventDateEnd = DateUtils.DateWithAddedMinutes(
-    eventDateStart,
-    EVENT_DURATION
-  );
-  const eventLocation = `בית החולים ${LocaleUtils.getHospitalName(
-    bookedAppointment.hospital
-  )}`;
-  const eventDescription = `זכרון מנחם - ${eventLocation} - ${eventDateStart.toLocaleDateString(
-    "he-He",
-    DateUtils.ShortDateFormat
-  )}`;
-
-  const getGoogleCalendarEvent = (): any => ({
-    text: `תרומת טרומבוציטים`,
-    location: eventLocation,
-    details: eventDescription,
-    dates: `${DateUtils.CleanIsoDate(eventDateStart)}/${DateUtils.CleanIsoDate(
-      eventDateEnd
-    )}`,
-  });
-
-  const getAppleCalendarEvent = (): any => ({
-    title: "תרומת טרומבוציטים",
-    location: eventLocation,
-    description: eventDescription,
-    startTime: eventDateStart.toString(),
-    endTime: eventDateEnd.toString(),
-  });
-
-  const addToGoogleCalendar = () => {
-    const googleEventUrl =
-      "https://calendar.google.com/calendar/u/0/r/eventedit?";
-    const eventParams = Object.entries(getGoogleCalendarEvent())
-      .map(([key, val]) => `${key}=${val}`)
-      .join("&");
-    window.open(googleEventUrl + eventParams, "_blank");
-  };
-
   return (
     <ZMScreen hasBurgerMenu>
       <div className={styles.pinkContainer}>
@@ -103,22 +63,14 @@ export default function UpcomingDonationScreen({
 
               <div className={styles.detailLabel}>מתי?</div>
               <div className={styles.detailValue}>
-                {eventDateStart.toLocaleDateString(
-                  "he-He",
-                  DateUtils.ShortDateFormat
-                )}
+                {new Date(
+                  bookedAppointment.donationStartTimeMillis
+                ).toLocaleDateString("he-He", DateUtils.ShortDateFormat)}
               </div>
-              {DeviceUtils.getMobileOperatingSystem() === "android" ? (
-                <div className={styles.link} onClick={addToGoogleCalendar}>
-                  הוספה ליומן
-                </div>
-              ) : (
-                <div className={styles.link}>
-                  <ICalendarLink event={getAppleCalendarEvent()}>
-                    הוספה ליומן
-                  </ICalendarLink>
-                </div>
-              )}
+              <EventCalendarLink
+                bookedAppointment={bookedAppointment}
+                eventDurationByMinutes={EVENT_DURATION_BY_MINUTES}
+              />
             </div>
           </div>
         </div>
